refactor: paths and imports
