+++ src/commands/brotli/compress.js
@@ -1,7 +1,7 @@
 import path from 'path';
 import store from '../../store/store.js';
 import { BROTLI_ACTIONS } from '../../constants/brotliActions.js';
-import { brotli } from './brotli.js';
+import { brotli } from '../../helpers';
 
 export const compress = async (filePath, destinationPath) => {
   const readFile = path.resolve(store.currentPath, filePath);
