+++ src/commands/brotli/index.js
@@ -1,0 +1,4 @@
+import { compress } from './compress.js';
+import { decompress } from './decompress.js';
+
+export { compress, decompress };
