+++ src/commands/files/index.js
@@ -1,0 +1,8 @@
+import { add } from './add.js';
+import { cat } from './cat.js';
+import { cp } from './cp.js';
+import { mv } from './mv.js';
+import { rm } from './rm.js';
+import { rn } from './rn.js';
+
+export { add, cat, cp, mv, rm, rn };
