+++ src/commands/hash/index.js
@@ -1,0 +1,3 @@
+import { hash } from './hash.js';
+
+export { hash };
