+++ src/commands/nwd/index.js
@@ -1,0 +1,5 @@
+import { cd } from './cd.js';
+import { ls } from './ls.js';
+import { up } from './up';
+
+export { cd, ls, up };
