+++ src/commands/os/arguments/architecture.js
