+++ src/commands/os/arguments/cpus.js
