+++ src/commands/os/arguments/index.js
@@ -1,0 +1,7 @@
+import { architecture } from './architecture.js';
+import { cpus } from './cpus.js';
+import { eol } from './eol.js';
+import { homedir } from './homedir.js';
+import { username } from './username.js';
+
+export { architecture, cpus, eol, homedir, username };
