+++ src/commands/os/index.js
@@ -1,0 +1,3 @@
+import { os } from './os.js';
+
+export { os };
