+++ src/commands/os/os.js
@@ -1,11 +1,7 @@
 import { stdin } from 'process';
-import { OS_ARGUMENTS } from '../../constants/osArguments.js';
+import { architecture, cpus, eol, homedir, username } from './arguments';
 
-import { eol } from './eol.js';
-import { cpus } from './cpus.js';
-import { homedir } from './homedir.js';
-import { username } from './username.js';
-import { architecture } from './architecture.js';
+import { OS_ARGUMENTS } from '../../constants/osArguments.js';
 
 export const os = (osArgument) => {
   switch (osArgument) {
